requirements.txt
# Core dependencies (pinned versions for reproducibility)
numpy
scipy
tifffile
PyYAML
joblib
